Container compaction: postgres-image
Updates bin and conf paths to combine all postgres subscripts into a single main directory.

Updates bin scripts to use a variable for the as the base path when grabbing scripts or conf files.

Updates base path from `/opt/cpm` to `/opt/crunchy`
- update bashrc and bash_profile to reference new path
- update bin scripts to reference new path

Adds common start.sh script to run in correct mode

Compacted into single postgres image:
- pgdump
- pgrestore
- pgbasebackup
- pgbench
- backup
- sqlrunner

The crunchy-postgres is now used as a base to build crunchy-postgres-ha and
crunchy-postgres-gis. The crunchy-postgres-ha image is used as the base to
crunchy-postgres-gis-ha. This means that each new layer is only adding layers
that are needed to augment the PostgreSQL base image. The layers have also been
updated so layers are not duplicated between the different dockerfiles. This
includes creating directories and volumes in the image and ensuring that the
necessary packages are included for each running mode.

The crunchy-postgres image can now run in different modes based on which
PostgreSQL based task needs to be completed. The modes correspond to the
compacted images and can be set using the `MODE` environment variable when
running the image. The start command for the compacted crunchy-postgres image is
now a start script that will run the correct start script based on how `MODE` is
set.

The directory structure in the repo is now updated so that the binary and config
files can be added to the compacted image in one layer of the image. The files
for the old images are now contained under a single directory
(`postgres_common`) that lives in both the `bin` and `conf` directories.
`postgres_common` contains the files needed for the `crunchy-postgres` image to
run in any supported mode.

The modules directory is now included in the postgres_common directory.
Previously some of these modules were dupicated in the crunchy-postgres and
crunchy-postgres-ha images. Now the duplicated module files have been
compacted where possible. They are included in the postgres image so that they
will also be available in the postgres-ha, postgres-gis, and postgres-gis-ha
images.

Update bin/postgres_common/sqlrunner/start.sh

Co-authored-by: Jonathan S. Katz <jkatz@users.noreply.github.com>